Move tryWrite smoke coverage into dedicated SmokeTest
